// src/lib/enrollment-schemas.ts
export const AddressSchema = z.object({
  logradouro: z.string().min(1, 'Logradouro é obrigatório').max(255),
  numero: z.string().min(1, 'Número é obrigatório').max(20),
  pontoReferencia: z.string().optional(),
  bairro: z.string().min(1, 'Bairro é obrigatório').max(255),
  municipio: z.string().min(1, 'Município é obrigatório').max(255),
  uf: z.string().min(2, 'UF é obrigatória').max(2),
  cep: z.string().refine((cep) => {
    const cleanCep = cep.replace(/\D/g, '')
    return cleanCep.length === 8
  }, 'CEP deve ter 8 dígitos'),
  telefoneResidencial: z.string().optional(),
  telefoneContato: z.string().optional()
})

export const DocumentsSchema = z.object({
  certidaoNascimento: z.string().optional(),
  municipioNascimento: z.string().optional(),
  municipioRegistro: z.string().optional(),
  cartorioRegistro: z.string().optional(),
  cpfCrianca: z.string().optional(),
  rgCrianca: z.string().optional(),
  dataEmissaoRg: z.date().optional(),
  orgaoEmissor: z.string().optional()
})

export const AddressDocumentsSchema = AddressSchema.merge(DocumentsSchema)
// ...
export type AddressDocuments = z.infer<typeof AddressDocumentsSchema>